Changing types to int
Changing NodeId back to int
Changing REAL to int, compiles, does not work

+++ submodular-flow/spd2.hpp
@@ -12,7 +12,7 @@
 
 class SubmodularPrimalDual2 {
     public:
-        typedef size_t NodeId;
+        typedef int NodeId;
         typedef size_t Label;
         typedef std::shared_ptr<Clique> CliquePtr;
         typedef std::vector<REAL> UnaryCost;
